Update Script locations
Move Player script

+++ Assets/Scripts/Shooting/Target.cs
@@ -6,7 +6,7 @@ public class Target : MonoBehaviour, IDamageable
 {
     [SerializeField] GameObject impactEffect;
 
-    public void Damage (RaycastHit hit, int damage)
+    public void Damage (RaycastHit hit, float damage)
     {
         GameObject impactParticle = Instantiate(impactEffect, hit.point, Quaternion.LookRotation(hit.normal));
         Destroy(impactParticle, 1f);
